// src/app/models/cliente.ts
export class Cliente {
    Id: number | undefined;
    cdCliente: number | undefined;
    nomeCliente: string | undefined;
    cnpj: string | undefined;
    escricaoEstadual: string | undefined;
    dtInclusao: string | undefined;
    categoria: number | undefined;
    cdSituacao: number | undefined;

}